requirements.txt
Flask==0.10.1
selenium==4.1.0
python-dotenv==0.17.1